Improve API docs descriptions
